restructure blockchain, dht and tests
