+++ src/external/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod bitcoin;
-pub(crate) mod time;
+pub mod bitcoin;
+pub mod time;
